set up nginx configs and make passenger happy
Passenger will only load the app if the bundler path is explicitly set.
We use /home/vagrant/.bundle so that it's strictly in the VM and doesn't
interfere with the app's working directly.

Passenger will also inherit an empty PATH from nginx, so the execjs gem
won't find the nodejs binary. We tell nginx to bring in the PATH env to
keep everybody happy.